requirements.txt
Django==2.2.12
django-cors-headers==3.7.0
django-import-export==2.2.0
djangorestframework==3.11.1
xlwt==1.3.0
Pillow==7.1.2
request==3.3.2